+++ editor/instancer/voxel_instancer_editor_plugin.h
@@ -30,6 +30,12 @@ private:
 	static void _bind_methods();
 
 	MenuButton *_menu_button = nullptr;
+	// TODO Use an ObjectID for referencing... it's a neverending struggle to keep checking pointer validity.
+	// When closing a scene while the node is selected, Godot will call `make_visible(false)` and `edit(null)` AFTER
+	// having deleted all the nodes, which means this plugin will be left with a dangling pointer when it's time to
+	// turn off the node's debug drawing feature... we could say "oh but you're not supposed to do that" or "you should
+	// use the gizmos system", but sorry, gizmos are too slow and complicated for this, and it is a feature available in
+	// game too.
 	VoxelInstancer *_node = nullptr;
 	VoxelInstancerStatView *_stat_view = nullptr;
 };