+++ terrain/instancing/voxel_instance_generator.h
@@ -2,6 +2,7 @@
 #define VOXEL_INSTANCE_GENERATOR_H
 
 //#include "../../storage/voxel_buffer.h"
+#include "../../generators/graph/voxel_graph_function.h"
 #include "../../util/godot/classes/noise.h"
 #include "../../util/math/transform3f.h"
 #include "../../util/math/vector3i.h"
@@ -108,6 +109,9 @@ public:
 	void set_noise(Ref<Noise> noise);
 	Ref<Noise> get_noise() const;
 
+	void set_noise_graph(Ref<pg::VoxelGraphFunction> func);
+	Ref<pg::VoxelGraphFunction> get_noise_graph() const;
+
 	void set_noise_dimension(Dimension dim);
 	Dimension get_noise_dimension() const;
 
@@ -122,8 +126,13 @@ public:
 		return int(x) | (int(y) << 1) | (int(z) << 2);
 	}
 
+#ifdef TOOLS_ENABLED
+	void get_configuration_warnings(PackedStringArray &warnings) const;
+#endif
+
 private:
 	void _on_noise_changed();
+	void _on_noise_graph_changed();
 
 	static void _bind_methods();
 
@@ -143,6 +152,19 @@ private:
 	Ref<Noise> _noise;
 	Dimension _noise_dimension = DIMENSION_3D;
 	float _noise_on_scale = 0.f;
+
+	// TODO Protect noise and noise graph members from multithreaded access
+
+	// Required inputs:
+	// - X
+	// - Y
+	// - Z
+	// Possible outputs:
+	// - Density
+	Ref<pg::VoxelGraphFunction> _noise_graph;
+	// TODO Sampling mode:
+	// - Per vertex: recommended if many items with high density need it (will be shared among them)
+	// - Per instance: recommended if items with low density need it
 
 	// Stored separately for editor
 	float _min_slope_degrees = 0.f;