+++ terrain/instancing/voxel_instance_library.cpp
@@ -118,6 +118,17 @@ void VoxelInstanceLibrary::remove_listener(IListener *listener) {
 	_listeners.erase(it);
 }
 
+void VoxelInstanceLibrary::get_configuration_warnings(PackedStringArray &warnings) const {
+	for (auto it = _items.begin(); it != _items.end(); ++it) {
+		Ref<VoxelInstanceLibraryItem> item = it->second;
+		ZN_ASSERT_CONTINUE(item.is_valid());
+		get_resource_configuration_warnings(**item, warnings,
+				[&it]() { //
+					return String("Item {0} (\"{1}\"): ").format(varray(it->first, it->second->get_item_name()));
+				});
+	}
+}
+
 bool VoxelInstanceLibrary::_set(const StringName &p_name, const Variant &p_value) {
 	const String property_name = p_name;