+++ tests/tests.cpp
@@ -76,6 +76,7 @@ void run_voxel_tests() {
 	VOXEL_TEST(test_voxel_graph_unused_single_texture_output);
 	VOXEL_TEST(test_voxel_graph_spots2d_optimized_execution_map);
 	VOXEL_TEST(test_voxel_graph_unused_inner_output);
+	VOXEL_TEST(test_voxel_graph_function_execute);
 	VOXEL_TEST(test_island_finder);
 	VOXEL_TEST(test_unordered_remove_if);
 	VOXEL_TEST(test_instance_data_serialization);