+++ doc/source/changelog.md
@@ -22,6 +22,9 @@ Semver is not yet in place, so each version can have breaking changes, although 
         - Added `Spots2D` and `Spots3D` nodes, optimized for generating "ore patches"
         - Added shader support for `FastNoiseGradient2D` and `FastNoiseGradient3D` nodes
         - Editor: reworked context menu to add nodes, similar to VisualShader. Now has search bar, tree view and node descriptions.
+    - `VoxelGraphFunction`:
+        - Editor: the graph will now get compiled while editing, which provides some checks
+        - Editor: I/Os are setup automatically when the graph is compiled by default. Manual setup might be exposed in the future if necessary.
     - `VoxelTerrain`:
         - Added `VoxelTerrainMultiplayerSynchronizer`, which simplifies replication using Godot's high-level multiplayer API
         - Added `is_area_meshed` as an alternative to `VoxelTool.is_area_editable` for games using mesh colliders
@@ -32,6 +35,8 @@ Semver is not yet in place, so each version can have breaking changes, although 
         - Added support for `paste`
     - `VoxelMesherCubes`:
         - Added helper function to convert an image into a 1-voxel thick "sprite mesh"
+    - `VoxelInstanceGenerator`:
+        - Added noise graph property, so instances can also be filtered with a custom VoxelGraphFunction
     - `VoxelInstanceLibrary`:
         - Added `get_all_item_ids()` to allow iterating over all items of a library
     - `VoxelLibraryMultiMeshItem `: